testapi/requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
backcall==0.1.0
certifi==2018.10.15
chardet==3.0.4
decorator==4.3.0
Django==2.1.2
django-filter==2.0.0
django-filters==0.2.1
djangorestframework==3.9.0
idna==2.7
ipdb==0.11
ipython==7.1.1
ipython-genutils==0.2.0
jedi==0.13.1
more-itertools==4.3.0
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
pluggy==0.8.0
prompt-toolkit==2.0.7
psycopg2==2.7.5
ptyprocess==0.6.0
py==1.7.0
Pygments==2.2.0
pytest==3.9.3
pytest-django==3.4.3
pytz==2018.7
requests==2.20.0
responses==0.10.2
six==1.11.0
traitlets==4.3.2
urllib3==1.24
wcwidth==0.1.7
